requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.14.0
generics==6.0.0
gunicorn==20.0.4
Pillow==9.4.0
psycopg==3.1.8
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
whitenoise==6.3.0
xlwt==1.3.0
